Remove validation of common options
Remove validation of common options and move the logic of validation outside of lib. This is required as when you pass option that's not common, but is still valid, validating the common options would cause process exit. Add validation of shorthands options in validateYargsArguments. Add validation for boolean options - if you pass value to a boolean known option, the process exits.

Remove validation of common options

Remove validation of common options and move the logic of validation outside of lib. This is required as when you pass option that's not common, but is still valid, validating the common options would cause process exit. Add validation of shorthands options in validateYargsArguments. Add validation for boolean options - if you pass value to a boolean known option, the process exits.