Fix setting of profile-dir
Add profileDir to knownOptions to prevent breaking of process when --profile-dir is passed on command line. The reason for breaking was yargs behavior, that adds values with dash as two keys - one with dash (profile-dir) and another one with removed dash and fist letter after it is replaced with its upper value.
Add setProfileDir function to options. It is used to set the profileDir based on client's default profile dir. The function adds the value to yargs.argv and to the exported options, so any code that requires options or yargs.argv will be able to use profile-dir.