Replace retryCount with timeout
When starting Android emulator we were using --retryCount option, which is difficult to be explained. It it replaced with --timeout option, which is the allowed time to start the emulator device(wait to boot). The default value is 120 seconds. retryCount option is removed as it is not used anymore.
When using var date = new Date(); and try to get date.getTime(), it returns the EPOCH time for the moment when new Date was called, not for the current time. Added helper function - getCurrentEpochTime which creates new Date instance and returns its Epoch value. Update android-emulator-services to use this function, so the operation will break correctly when the time for starting emulator passes the set timeout.
Update error messages.