Validate iOS appstore commands
Use iTunes validation for appstore commands. Add IiTunesValidator interface with a single method - getError. Chang ITunesValidator class to implement this interface and move the class to new file. Use the implementation to validate appstore list and appstore upload commands.
The validator is required when you are using x64 bit Node on Windows - in this case we are unable to use iTunes as it doesn't have x64 bit dlls.