Use genymotion emulator
Add geny option. Add check if it is specified to start the player with geny option and use player from Genymotion folder. Add check dependency for Genymotion if options.geny is specified.

Check if there's already a running emulator with specified name. For genymotion emulator:
 - for genymotion -> "adb shell -s <emulator id> getprop ro.product.model". I use this call for each running emulator and check if one of them is the required for starting.
 - for android emulator -> write on the emulator port (specified in emulator-<port_num> when listed with `adb devices`) command "avd name" and parse the result.
If there's already a started emulator with this image, we just have to deploy the changes on it and do not start another one (as we've done until know when --avd option is specified).
Add geny option. Add check if it is specified to start the player with geny option and use player from Genymotion folder. Add check dependency for Genymotion if options.geny is specified.
Add check if --avd and --geny options are both specified and throws error in such case.

http://teampulse.telerik.com/view#item/274611