Add check if 7za can be executed
If user cannot use 7zip, he is unable to create projects, unable to build etc. So on post-install we check if we are able to execute the 7za file and if ENOENT is return we show the user where he can find system requirements.

Fix host-info paths in some files and remove host-info reference from some other files.

http://teampulse.telerik.com/view#item/285027