Add dashedOptions for each command
Each command can use specific dashed options (`-- avd <name>` for example is valid for one command, but not for other). Add dashedOptions parameter to ICommand interface and make sure to use it for validation when it exists. In case there's no dashedOptions paramter, use the common options.

Use correct types for yargs - "string", "boolean", etc. instead of 0, 1, 2 which were the result of our enum. Replace the enum with class and remove code, which we do not need anymore as yargs is doing its work as expected.
After yargs is updated to latest version, some fixes have to be applied - now when `yargs.options(opts).argv` is passed, it returns all members of `opts`, but the ones that are not part of process.argv have value undefined.