Provide a way to expose Promisified public API
Use Typescript decorators in order to provide an easy way to expose public API.
Created two decorators:
* `register` - class decorator that can be used instead of `$injector.register`. Usage is `@decorators.register(<module_name>)`
* `promisify` - creates a wrapper around method in order to return promise. The method itself is kept untouched, but the wrapped is added to the publicApi object. Usage is `@decorators.promisify(<module_name>)`

Add requirePublic method in yok in order to create public api. When a module has some methods that must be added to public API, it has to be required with "requirePublic".

Set `getFileSize` of fs module to be promisified in order to allow testing.