Introduce messages-service
Used for extracting messages out of given .json files. If given a key be it simple (Some_error_message) or complex (iOS.iTunes.Connection_Refused_Error) should return a message if one is found in any given file. If no value is found returns the key itself. Given .json files are prioritized - that is to say if a key is found in the first file it is immediately returned. That way messages can be overriden - for example appbuilder.messages.json may override keys from common.messages.json