Introduce `dev-generate-messages` command
That way an interface (IMessages) and an implementation file is generated which can be used for autocompletion when accessing JSON-predefined error messages.
Includes a bit of refactoring - extracting some classes from swagger into common.
Usage of the new messages service can be seen in `devices-services` class
I'm committing the first draft of messages.d.ts and messages.ts which will hopefully grow in the future.