Allow defining subcommands of subcommands
In case you have command called 'prop|print' and you want to add subcommand to it,
the injector will not try to resolve it. For example 'prop|print|myprop' will be resolved
as command 'prop|print' with argument 'myprop'.

Fix this by executing the full `_.each` and adjust the real parameters based on the last matching one.
Add unit tests.