Update DEVICE_LOOKING_INTERVAL to 2200
Checking for currently attached iOS devices is taking 1 second. Also we are spawning check for iOS devices on each second.
So we end up with always checking for iOS devices which is blocking operation. This way everything is working really slow.
Set the interval for device checking to 2200.
Set version to 0.0.4