Do NOT execute ios device commands if iTunes error present
Instead of directly trying to execute ios device-related commands, first check for an error in iTunes.
If an error is present cache it and do not execute anything anymore.
If no error is present try and execute command but do NOT cache. Next command execution attempt will trigger a new check to make sure everything's still all right