Introduce proton logger
Fix `EPERM: operation not permitted, write` error when started from Proton installation and trying to write to stdout.
The process does not have stdout, so all logger calls will cause such failure.
Introduce new logger, special for Proton, that will not do anything at the moment. We'll extend it in the future to send messages to Proton with the messages.