Fix deploy on iOS devices
When we deploy on iOS devices, the first sent package sometimes receives error code 21. It turned out this error is not breaking the deployment, so ignore it once and continue deploying.
The error occurs only when we have parallel futures executing for device detection.

Also when error is raised during application starting, we reject the Promise for deploy on device. This is not correct, so ignore such errors and resolve the promise.