Add default compilation for TypeScript files
Currently TypeScript projects are compiled by enumerating their files and spawning tsc with specific options.
We do not need this when tsconfig.json exists as the TypeScript compiler uses it directly.
Add new method to TypeScriptCompilationService in order to omit our own options passed to tsc.
Remove initialize method as it is not used anymore.