Fix uncaughtExceptionListener for our unit tests
When staticConfig is not registered (in our tests), when uncaught exception is raised, the listener fails with error that staticConfig cannot be resolved.
This makes it hard for debugging. Try-catch the resolve of staticConfig and set the disalbeAnalytics to true when staticConfig cannot be resolved.
This way the tests will not have to register analyticsService.