Add unit tests for AnalyticsService
Move unit tests for AnalyticsService from icenium-cli to common.

Added new Tests:
* setStatus
  * tracks that user had disabled feature tracking
  * tracks that user had enabled feature tracking
  * tracks that user had disabled exceptions tracking
  * tracks that user had enabled exceptions tracking

* checkConsent
  * sends information that user had accepted feature tracking
  * sends information that user had rejected feature tracking

Also remove the isEverythingDisabled check from analytics start method.
In case featureUsageTracking is disabled and exceptions reporting is disabled,
we might still want to track some information, so make sure eqatecMonitor can start.
Such case is when ExceptionsTracking is disabled, we ask the user if he wants to enable FeatureTracking and he refuses.
All tracking is disabled, but we want to track down that the user refused to enable feature tracking.