Unmute stdout after prompter execution
In order to prevent double printing of stdout data, CLI mutes the stdout. However this is done in prompter constructor, so every class, that depends on the prompter, automatically mutes the stdout.
Basically CLI mutes the stdout on almost all commands. This causes issues when `tns plugin add` command is used and the plugin wants to prompt the user for something on postinstall.
Fix this by calling mute exactly before prompting the user and calling unmute immediately after user sends his answer.