Fix starting many adb processes
When new device is attached, common lib is trying to get all of its applications and to send intents for each of them in order to check is livesync supported for them.
This leads to tens or even hundreds of started adb processes. At some point, when you have more than one device, adb cannot handle the whole flow and it suicides itself.
After that you have to kill manually the only existing adb process or it will never detect any devices.
In order to prevent this behavior, change the way of detection is LiveSync supported. Expose new public method that will be used to check if LiveSync is supported for specific devices (and one appIdentifier).
Remove the reporting of isLiveSyncSupported in the "applicationInstalled" event.