Fix Android device detection when adb is not in the PATH
When adb is not in the PATH, common lib is trying to spawn it's private adb, which leads to recursion errors.
Make sure to spawn our own adb in this cases.
Also prevent shelljs cp errors by skipping new copy of files in case they are already there.
Prevent shelljs.chmod errors on Windows - we do not need to chmod on Windows.