Allow starting of NativeScript apps on Android via Proton
Currently when deploying NativeScript apps on Android via Proton, we are unable to start the application, as we pass incorrect StartPackageActivity.
Fix this by extracting the correct names in constants. When deploy is called from Proton side, we'll set the correct name in the `$project`.
After that the execution of start method will use the value from `$staticConfig`. StaticConfig will return the value of project's property.

`deployOnDevices` method now has additional argument - framework. If it is passed and the provided value is NativeScript, we'll use the NativeScript's StartPackageActivity.
In all other cases we'll use Cordova's StartPackageActivity.