Set iOS Device status to Unreachable only in constructor
Set iOS Device status to Unreachable only in constructor, not when some operation fails.
Most of the operations may fail a single time only, so the device is not Unreachable.
When device is trusted, we will remove current instance of iOS Device and we'll create new one, where the status will be Connected.