Stop device detection and checking for application updates when ops in progress
When you have deploy operation or livesync operation, we must stop device detection and checking for application changes as it breaks the deployment on iOS Devices.
The problem is that in case ANY async callback is called during uploading .ipa file to iOS Device for installation, the installation fails with different errors (PackageExtractionFailed, PackageVerificationFailed, etc.).
Add new future inside the setInterval and in case it's not resolved, the next time when setInterval is called, it will see that there's a future pending and will not schedule new execution.
Also when stopDeviceDetectionInterval is called and the future from setInterval is still pending, wait for it before starting the real operation.