Add more unit tests for decorators
Add more unit tests for decorators and fix some issues that raised during writing tests.
Add `.catch` in the `exportedPromise` decorator in order to prevent "Unhandled Rejection Error" when one of the promises is rejected.
This will not break the current API behavior as the rejected promise will still be returned as result and the consumer of the API can add it's own catch.

The `exported` decorator was incorrectly using the `target` instead of the object from injector.

Update mocha-fibers dependency, so we can use the done callback.