Allow debugging of WebViews inside NativeScript applications
When there are webviews inside NativeScript application, mobile-cli-lib does not return them in getDebuggableApps as they have the same appIdentifier.
Also when trying to map abstract port for NativeScript application with WebViews, mobile-cli-lib finds the WebView first and maps port for it. In order to fix this, add `framework` parameter that will be passed from Proton.