Add method to transpile files in the public api
We need to add method to transpile typescript files to the public api.
Also in the typeScriptService we need to use npm binary to get the typescript module from npm.
The exposed method will transpile all files in the project or only the files which are passed as parameter.
Compiler options can be specified as parameter. If they are not specified the method will search for tsconfig.json file and get them from it. If there is no tsconfig.json file the method will use the default options.